backend/requirements/base.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.3

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1

# Cache
redis==5.0.1
django-redis==5.4.0

# Utilities
python-decouple==3.8
Pillow==10.1.0
pytz==2023.3

# API Documentation
drf-spectacular==0.26.5
